Moved checks_with_crest into zmat/base/extra to try and remove circular import
